+++ v2/src/modules/address/ModuleAddressRecentNfts.vue
@@ -60,8 +60,8 @@ import { useGetOwnersNftTokensQuery } from './apollo/AddressTokens/tokens.genera
 import { useRouter } from 'vue-router'
 import { ADDRESS_ROUTE_QUERY, ROUTE_NAME } from '@core/router/routesNames'
 import AppNoResult from '@/core/components/AppNoResult.vue'
-import TokenNftImg from '../tokens/components/TokenNFT/TokenNftImg.vue'
-import { NFTDetails } from '../tokens/components/TokenNFT/propModel'
+import TokenNftImg from '../tokens/components/token-nft/TokenNftImg.vue'
+import { NFTDetails } from '../tokens/components/token-nft/propModel'
 import { useGetNftsMeta } from '@core/composables/NftMeta/useGetNftsMeta.composable'
 import { NftId, generateId, generateMapId } from '@/core/composables/NftMeta/helpers'
 import { NftType } from '@/apollo/types'
