+++ v2/src/modules/search/ModuleAppSearch.vue
@@ -157,7 +157,7 @@ import { Buffer } from 'buffer'
 import { useCoinData } from '@core/composables/CoinData/coinData.composable'
 import { formatUsdValue, formatNumber } from '@/core/helper/number-format-helper'
 import BN from 'bignumber.js'
-import { useGetTokenInfoByContractQuery } from '@module/tokens/apollo/TokenDetails/tokenDetails.generated'
+import { useGetTokenInfoByContractQuery } from '@module/tokens/apollo/token-details/tokenDetails.generated'
 import { useResolveName } from '@/core/composables/ResolveName/useResolveName'
 import { useStore } from '@/store'
 import { PortfolioItem } from '@/store/helpers'
