+++ v2/src/modules/tokens/ModuleAddFavToken.vue
@@ -38,7 +38,7 @@ import AppBtn from '@core/components/AppBtn.vue'
 import AppDialog from '@core/components/AppDialog.vue'
 import AppNoResult from '@core/components/AppNoResult.vue'
 import AppInput from '@core/components/AppInput.vue'
-import TokenMarketInfoTableRow from '@module/tokens/components/TokenMarketInfo/TableRowTokenMarketInfo.vue'
+import TokenMarketInfoTableRow from '@module/tokens/components/token-market-info/TableRowTokenMarketInfo.vue'
 import { computed, reactive } from 'vue'
 import { useStore } from '@/store'
 import { useCoinData } from '@core/composables/CoinData/coinData.composable'
