+++ v2/src/modules/tokens/ModuleTokenDetails.vue
@@ -108,8 +108,8 @@ import { reactive, computed, onMounted, watch } from 'vue'
 import AppAdButtonsLarge from '@/core/components/AppAdButtonsLarge.vue'
 import AppBtnIcon from '@core/components/AppBtnIcon.vue'
 import AppCopyToClip from '@/core/components/AppCopyToClip.vue'
-import TokenDetailsErc20 from '@module/tokens/components/TokenDetails/TokenDetailsERC20.vue'
-import TokenDetailsNft from '@module/tokens/components/TokenDetails/TokenDetailsNFT.vue'
+import TokenDetailsErc20 from '@module/tokens/components/token-details/TokenDetailsERC20.vue'
+import TokenDetailsNft from '@module/tokens/components/token-details/TokenDetailsNFT.vue'
 import TokenTransfers from '@module/tokens/components/TokenTransfers.vue'
 import TokenHolders from '@module/tokens/components/TokenHolders.vue'
 import AppTabs from '@/core/components/AppTabs.vue'
@@ -121,7 +121,7 @@ import {
     GetTokenInfoByContractQuery,
     useGetNftContractMetaQuery,
     NftCollectionFragment
-} from '@module/tokens/apollo/TokenDetails/tokenDetails.generated'
+} from '@module/tokens/apollo/token-details/tokenDetails.generated'
 import { eth } from '@core/helper'
 import { ErrorMessageToken } from '@module/tokens/models/ErrorMessagesForTokens'
 import { Q_TOKEN_DETAILS } from '@core/router/routesNames'
