+++ v2/src/modules/tokens/ModuleTokensInfo.vue
@@ -130,7 +130,7 @@ import AppNoResult from '@core/components/AppNoResult.vue'
 import AppPagination from '@core/components/AppPagination.vue'
 import AppMenu from '@core/components/AppMenu.vue'
 import { Tab } from '@/core/components/props'
-import TokenMarketInfoTableRow from '@module/tokens/components/TokenMarketInfo/TableRowTokenMarketInfo.vue'
+import TokenMarketInfoTableRow from '@module/tokens/components/token-market-info/TableRowTokenMarketInfo.vue'
 import ModuleAddFavToken from './ModuleAddFavToken.vue'
 import { computed, reactive } from 'vue'
 import { useStore } from '@/store'
