+++ v2/src/modules/tokens/apollo/token-details-holder/tokenHolders.graphql
