+++ v2/src/modules/tokens/apollo/token-details-transfer/tokenTransfers.generated.ts
