+++ v2/src/modules/tokens/apollo/token-details/tokenDetails.generated.ts
