+++ v2/src/modules/tokens/components/token-details-transfer/TokenTransferTableRow.vue
@@ -114,7 +114,7 @@ import {
     TokenTransferFragment,
     Erc721TransferFragment,
     Erc1155TokenTransferFragment
-} from '@module/tokens/apollo/TokenDetailsTransfer/tokenTransfers.generated'
+} from '@module/tokens/apollo/token-details-transfer/tokenTransfers.generated'
 import BN from 'bignumber.js'
 import { computed, ref } from 'vue'
 import { formatFloatingPointValue, FormattedNumber } from '@core/helper/number-format-helper'
@@ -122,8 +122,8 @@ import { useDisplay } from 'vuetify'
 import { eth, timeAgo } from '@core/helper'
 import { TransferType } from '@/apollo/types'
 import { NftMetaFragment } from '@core/composables/NftMeta/nftMeta.generated'
-import TokenNftImg from '@module/tokens/components/TokenNFT/TokenNftImg.vue'
-import { NFTDetails } from '@module/tokens/components/TokenNFT/propModel'
+import TokenNftImg from '@module/tokens/components/token-nft/TokenNftImg.vue'
+import { NFTDetails } from '@module/tokens/components/token-nft/propModel'
 import Web3Utils from 'web3-utils'
 
 const { lgAndUp } = useDisplay()
