+++ v2/src/modules/tokens/components/token-details/TokenDetailsCollectionImg.vue
