+++ v2/src/modules/tokens/components/token-details/TokenDetailsERC20.vue
@@ -45,7 +45,7 @@ import AppTokenIcon from '@/core/components/AppTokenIcon.vue'
 import BN from 'bignumber.js'
 import { formatPercentageValue, formatUsdValue } from '@core/helper/number-format-helper'
 import { useCoinData } from '@core/composables/CoinData/coinData.composable'
-import { TokenDetailsFragment as TokenInfo } from '@module/tokens/apollo/TokenDetails/tokenDetails.generated'
+import { TokenDetailsFragment as TokenInfo } from '@module/tokens/apollo/token-details/tokenDetails.generated'
 import { ErrorMessageToken } from '@module/tokens/models/ErrorMessagesForTokens'
 import { computed } from 'vue'
 import { MarketDataFragment as TokenMarketData } from '@/core/composables/CoinData/getLatestPrices.generated'
