+++ v2/src/modules/tokens/components/token-market-info/TableRowTokenMarketInfo.vue
