+++ v2/src/modules/tokens/components/token-nft/TokenNftDetails.vue
