+++ v2/src/modules/txs/components/TxActionRow.vue
@@ -59,7 +59,7 @@ import AppTransformHash from '@core/components/AppTransformHash.vue'
 import AppTableRow from '@core/components/AppTableRow.vue'
 import AppTokenIcon from '@core/components/AppTokenIcon.vue'
 import AppTooltip from '@/core/components/AppTooltip.vue'
-import TokenNftImg from '@module/tokens/components/TokenNFT/TokenNftImg.vue'
+import TokenNftImg from '@module/tokens/components/token-nft/TokenNftImg.vue'
 import { computed, PropType, ref } from 'vue'
 import { Action } from '../types'
 import { eth } from '@core/helper'
@@ -70,7 +70,7 @@ import { useDisplay } from 'vuetify/lib/framework.mjs'
 import { formatNonVariableEthValue, FormattedNumber, formatFloatingPointValue } from '@/core/helper/number-format-helper'
 import { MarketDataFragment } from '@core/composables/CoinData/getLatestPrices.generated'
 import { NftMetaFragment } from '@core/composables/NftMeta/nftMeta.generated'
-import { NFTDetails } from '@module/tokens/components/TokenNFT/propModel'
+import { NFTDetails } from '@module/tokens/components/token-nft/propModel'
 import BigNumber from 'bignumber.js'
 import { hexToNumberString } from 'web3-utils'
 import { useI18n } from 'vue-i18n'
